12/2 updated code
now have some semblance of an edge detected image, though the output may not be in black and white because for some reason greyscale is not greyscaling. should check in with wes about this, and also out edge detection might be different? from the energy maps that they are talking about online, so check in with him about that too maybe